Handle insert of each key in a separate transaction
Currently, a call to Insert results in a single database transaction for all
of the keys being inserted. In the case of a hockeypuck-load, this can mean
that one transaction is used for thousands of keys. If one of these keys fails
to insert (for example, the keywords exceeds the maximum size of a tsvector
field), then every key following that will fail to insert since the transaction
is now invalid.

Handling keys on a transaction per key basis means that a single key insert
failure does not result in multiple failures. Additionally, this requires
more consistent resources from PostgreSQL.

While here, cap the number of errors that can occur, since if hundres or
thousands of errors are occurring something is going fairly wrong and it
just makes it harder to figure out what the real problem is.

Also add a missing error check for a json.Marshal call.